worker/requirements.txt
pika==1.3.2
Pillow==10.4.0